Rearrange the `ops` module
