+++ src/ops/draw.rs
@@ -1,5 +1,4 @@
 use libc;
-use std::sync::Arc;
 use std::sync::mpsc::channel;
 
 use Display;
@@ -7,10 +6,9 @@ use Display;
 use fbo::{self, FramebufferAttachments};
 
 use uniforms::{Uniforms, UniformValue, SamplerBehavior};
-use {DisplayImpl, Program, DrawParameters, Rect, Surface, GlObject, ToGlEnum};
+use {Program, DrawParameters, GlObject, ToGlEnum};
 use index_buffer::IndicesSource;
 use vertex_buffer::VerticesSource;
-use texture;
 
 use {program, vertex_array_object};
 use {gl, context};
@@ -127,216 +125,6 @@ pub fn draw<'a, I, U>(display: &Display,
     if let Some(rx) = rx {
         rx.recv().unwrap();
     }
-}
-
-pub fn clear_color(display: &Arc<DisplayImpl>, framebuffer: Option<&FramebufferAttachments>,
-    red: f32, green: f32, blue: f32, alpha: f32)
-{
-    let fbo_id = display.framebuffer_objects.as_ref().unwrap()
-                        .get_framebuffer_for_drawing(framebuffer, &display.context);
-
-    let (red, green, blue, alpha) = (
-        red as gl::types::GLclampf,
-        green as gl::types::GLclampf,
-        blue as gl::types::GLclampf,
-        alpha as gl::types::GLclampf
-    );
-
-    display.context.exec(move |: mut ctxt| {
-        fbo::bind_framebuffer(&mut ctxt, fbo_id, true, false);
-
-        unsafe {
-            if ctxt.state.clear_color != (red, green, blue, alpha) {
-                ctxt.gl.ClearColor(red, green, blue, alpha);
-                ctxt.state.clear_color = (red, green, blue, alpha);
-            }
-
-            ctxt.gl.Clear(gl::COLOR_BUFFER_BIT);
-        }
-    });
-}
-
-pub fn clear_depth(display: &Arc<DisplayImpl>, framebuffer: Option<&FramebufferAttachments>,
-    value: f32)
-{
-    let value = value as gl::types::GLclampf;
-    
-    let fbo_id = display.framebuffer_objects.as_ref().unwrap()
-                        .get_framebuffer_for_drawing(framebuffer, &display.context);
-
-    display.context.exec(move |: mut ctxt| {
-        fbo::bind_framebuffer(&mut ctxt, fbo_id, true, false);
-
-        unsafe {
-            if ctxt.state.clear_depth != value {
-                ctxt.gl.ClearDepth(value as f64);        // TODO: find out why this needs "as"
-                ctxt.state.clear_depth = value;
-            }
-
-            ctxt.gl.Clear(gl::DEPTH_BUFFER_BIT);
-        }
-    });
-}
-
-pub fn clear_stencil(display: &Arc<DisplayImpl>, framebuffer: Option<&FramebufferAttachments>,
-    value: int)
-{
-    let value = value as gl::types::GLint;
-
-    let fbo_id = display.framebuffer_objects.as_ref().unwrap()
-                        .get_framebuffer_for_drawing(framebuffer, &display.context);
-
-    display.context.exec(move |: mut ctxt| {
-        fbo::bind_framebuffer(&mut ctxt, fbo_id, true, false);
-
-        unsafe {
-            if ctxt.state.clear_stencil != value {
-                ctxt.gl.ClearStencil(value);
-                ctxt.state.clear_stencil = value;
-            }
-
-            ctxt.gl.Clear(gl::STENCIL_BUFFER_BIT);
-        }
-    });
-}
-
-pub fn blit<S1: Surface, S2: Surface>(source: &S1, target: &S2, mask: gl::types::GLbitfield,
-    src_rect: &Rect, target_rect: &Rect, filter: gl::types::GLenum)
-{
-    let ::BlitHelper(display, source) = source.get_blit_helper();
-    let ::BlitHelper(_, target) = target.get_blit_helper();
-
-    let src_rect = src_rect.clone();
-    let target_rect = target_rect.clone();
-
-    // FIXME: we don't draw on it
-    let source = display.framebuffer_objects.as_ref().unwrap()
-                        .get_framebuffer_for_drawing(source, &display.context);
-    let target = display.framebuffer_objects.as_ref().unwrap()
-                        .get_framebuffer_for_drawing(target, &display.context);
-
-    display.context.exec(move |: ctxt| {
-        unsafe {
-            // trying to do a named blit if possible
-            if ctxt.version >= &context::GlVersion(4, 5) {
-                ctxt.gl.BlitNamedFramebuffer(source, target,
-                    src_rect.left as gl::types::GLint,
-                    src_rect.bottom as gl::types::GLint,
-                    (src_rect.left + src_rect.width) as gl::types::GLint,
-                    (src_rect.bottom + src_rect.height) as gl::types::GLint,
-                    target_rect.left as gl::types::GLint, target_rect.bottom as gl::types::GLint,
-                    (target_rect.left + target_rect.width) as gl::types::GLint,
-                    (target_rect.bottom + target_rect.height) as gl::types::GLint, mask, filter);
-
-                return;
-            }
-
-            // binding source framebuffer
-            if ctxt.state.read_framebuffer != source {
-                if ctxt.version >= &context::GlVersion(3, 0) {
-                    ctxt.gl.BindFramebuffer(gl::READ_FRAMEBUFFER, source);
-                    ctxt.state.read_framebuffer = source;
-
-                } else {
-                    ctxt.gl.BindFramebufferEXT(gl::READ_FRAMEBUFFER_EXT, source);
-                    ctxt.state.read_framebuffer = source;
-                }
-            }
-
-            // binding target framebuffer
-            if ctxt.state.draw_framebuffer != target {
-                if ctxt.version >= &context::GlVersion(3, 0) {
-                    ctxt.gl.BindFramebuffer(gl::DRAW_FRAMEBUFFER, target);
-                    ctxt.state.draw_framebuffer = target;
-
-                } else {
-                    ctxt.gl.BindFramebufferEXT(gl::DRAW_FRAMEBUFFER_EXT, target);
-                    ctxt.state.draw_framebuffer = target;
-                }
-            }
-
-            // doing the blit
-            if ctxt.version >= &context::GlVersion(3, 0) {
-                ctxt.gl.BlitFramebuffer(src_rect.left as gl::types::GLint,
-                    src_rect.bottom as gl::types::GLint,
-                    (src_rect.left + src_rect.width) as gl::types::GLint,
-                    (src_rect.bottom + src_rect.height) as gl::types::GLint,
-                    target_rect.left as gl::types::GLint, target_rect.bottom as gl::types::GLint,
-                    (target_rect.left + target_rect.width) as gl::types::GLint,
-                    (target_rect.bottom + target_rect.height) as gl::types::GLint, mask, filter);
-
-            } else {
-                ctxt.gl.BlitFramebufferEXT(src_rect.left as gl::types::GLint,
-                    src_rect.bottom as gl::types::GLint,
-                    (src_rect.left + src_rect.width) as gl::types::GLint,
-                    (src_rect.bottom + src_rect.height) as gl::types::GLint,
-                    target_rect.left as gl::types::GLint, target_rect.bottom as gl::types::GLint,
-                    (target_rect.left + target_rect.width) as gl::types::GLint,
-                    (target_rect.bottom + target_rect.height) as gl::types::GLint, mask, filter);
-            }
-        }
-    });
-}
-
-pub fn read_attachment<P, T>(attachment: &fbo::Attachment, dimensions: (u32, u32),
-                             display: &Display) -> T          // TODO: remove Clone for P
-                             where P: texture::PixelValue + Clone + Send,
-                             T: texture::Texture2dData<Data = P>
-{
-    let (fbo, atch) = display.context.framebuffer_objects.as_ref().unwrap()
-                             .get_framebuffer_for_reading(attachment, &display.context.context);
-    read_impl(fbo, atch, dimensions, &display.context.context)
-}
-
-pub fn read_from_default_fb<P, T>(attachment: gl::types::GLenum, display: &Display) -> T          // TODO: remove Clone for P
-                                  where P: texture::PixelValue + Clone + Send,
-                                  T: texture::Texture2dData<Data = P>
-{
-    let (w, h) = display.get_framebuffer_dimensions();
-    let (w, h) = (w as u32, h as u32);      // TODO: remove this conversion
-    read_impl(0, attachment, (w, h), &display.context.context)
-}
-
-fn read_impl<P, T>(fbo: gl::types::GLuint, readbuffer: gl::types::GLenum,
-                   dimensions: (u32, u32), context: &context::Context) -> T          // TODO: remove Clone for P
-                   where P: texture::PixelValue + Clone + Send,
-                   T: texture::Texture2dData<Data = P>
-{
-    use std::mem;
-
-    let pixels_count = dimensions.0 * dimensions.1;
-
-    let pixels_size = texture::Texture2dData::get_format(None::<T>).get_size();
-    let (format, gltype) = texture::Texture2dData::get_format(None::<T>).to_gl_enum();
-
-    let (tx, rx) = channel();
-    context.exec(move |: mut ctxt| {
-        unsafe {
-            // binding framebuffer
-            fbo::bind_framebuffer(&mut ctxt, fbo, false, true);
-
-            // adjusting glReadBuffer
-            ctxt.gl.ReadBuffer(readbuffer);
-
-            // adjusting data alignement
-            if ctxt.state.pixel_store_pack_alignment != 1 {
-                ctxt.state.pixel_store_pack_alignment = 1;
-                ctxt.gl.PixelStorei(gl::PACK_ALIGNMENT, 1);
-            }
-
-            // reading
-            let total_data_size = pixels_count  as usize * pixels_size / mem::size_of::<P>();
-            let mut data: Vec<P> = Vec::with_capacity(total_data_size);
-            ctxt.gl.ReadPixels(0, 0, dimensions.0 as gl::types::GLint,
-                               dimensions.1 as gl::types::GLint, format, gltype,
-                               data.as_mut_ptr() as *mut libc::c_void);
-            data.set_len(total_data_size);
-            tx.send(data).ok();
-        }
-    });
-
-    let data = rx.recv().unwrap();
-    texture::Texture2dData::from_vec(data, dimensions.0 as u32)
 }
 
 // TODO: we use a `Fn` instead of `FnOnce` because of that "std::thunk" issue
