+++ src/ops/mod.rs
@@ -1,0 +1,9 @@
+pub use self::blit::blit;
+pub use self::clear::{clear_color, clear_depth, clear_stencil};
+pub use self::draw::draw;
+pub use self::read::{read_attachment, read_from_default_fb};
+
+mod blit;
+mod clear;
+mod draw;
+mod read;
